wifi-access-point-geolocation/requirements.txt
fastapi
uvicorn
requests
pymemcache
structlog
tenacity
python-dotenv
